Fixed formatting issues (patch by: Joe Orton)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@142538 c90b9560-bf6c-de11-be94-00142212c4b1

+++ sapi/cgi/cgi_main.c
@@ -923,7 +923,7 @@ int main(int argc, char *argv[])
 	char *script_file=NULL;
 	zend_llist global_vars;
 #if FORCE_CGI_REDIRECT
-	int force_redirect = 1;
+	long force_redirect = 1;
 	char *redirect_status_env = NULL;
 #endif
